Refactor: created user controller and router, update frontend routes, fix ExploreScreen when no results
